refactor: move slack bot implementation
